// src/lib/enums/bulk-field-type.enum.ts
export enum BulkFieldType {

    TEXT = 'text',

    NUMBER = 'number',

    EMAIL = 'email',

    PHONE = 'phone',

    DATE = 'date',

    DATETIME = 'datetime',

    TIME = 'time',

    SINGLE_SELECT = 'single-select',

    MULTI_SELECT = 'multi-select',

    CHECKBOX = 'checkbox',

    TEXTAREA = 'textarea',

    FILE = 'file',

    CUSTOM = 'custom',

    PASSWORD = "password"

}